+++ src/SimulateC.cpp
@@ -12,6 +12,12 @@
 //' @param T T system matrix of the State Space model component.
 //' @param R R system matrix of the State Space model component.
 //' @param Q Q system matrix of the State Space model component.
+//' @param P_star P_star system matrix of the State Space model component.
+//'   Only used when draw_initial = true.
+//' @param draw_initial Boolean indicating whether the initial state
+//'   should be simulated.
+//' @param eta_only Boolean indicating whether only the
+//'   disturbances should be simulated.
 //'
 //' @noRd
 // [[Rcpp::export]]
@@ -22,11 +28,14 @@ Rcpp::List SimulateC(const int& nsim,
                      const arma::cube& Z,
                      const arma::cube& T,
                      const arma::cube& R,
-                     const arma::cube& Q) {
+                     const arma::cube& Q,
+                     const arma::mat& P_star,
+                     const bool& draw_initial,
+                     const bool& eta_only) {
 
   // Number of dependent variables, state parameters and state disturbances
   int p = Z.n_rows, m = a.n_rows, r = Q.n_rows, r_tot = r * repeat_Q,
-      nsim_rep = nsim * repeat_Q, total = r_tot * nsim;
+      nsim_rep = nsim * repeat_Q, total = r_tot * nsim, m_nsim = m * nsim;
 
   // The last time point
   int N_min1 = N - 1;
@@ -45,15 +54,24 @@ Rcpp::List SimulateC(const int& nsim,
   // Initialise simulated disturbances, state, and dependent
   arma::cube eta(N, r_tot, nsim), a_cube(N, m, nsim), y(N, p, nsim);
   arma::mat eta_sim(r, nsim_rep), eta_temp(r_tot, nsim);
-  arma::mat a_temp = arma::repmat(a, 1, nsim);
-  a_cube.row(0) = a_temp;
 
   // Initialise vector of random draws
-  Rcpp::NumericVector draw(total);
+  Rcpp::NumericVector draw(total), a_draw(m_nsim);
 
-  // Helpers for calculating root of Q
-  arma::mat Q_root(r, r), U_Q(r, r), V_Q(r, r);
-  arma::colvec s_Q(r);
+  // Helpers for calculating root of Q and P_star
+  arma::mat Q_root(r, r), U_Q(r, r), V_Q(r, r),
+            P_root(m, m), U_P(m, m), V_P(m, m);
+  arma::colvec s_Q(r), s_P(m);
+
+  // Initial state
+  arma::mat a_temp = arma::repmat(a, 1, nsim);
+  if (draw_initial) {
+    arma::svd(U_P, s_P, V_P, P_star);
+    P_root = U_P * arma::diagmat(arma::sqrt(s_P)) * U_P.t();
+    a_draw = Rcpp::rnorm(m_nsim);
+    a_temp = a_temp + P_root * arma::mat(a_draw.begin(), m, nsim);
+  }
+  a_cube.row(0) = a_temp;
 
   // Initial root of Q
   arma::svd(U_Q, s_Q, V_Q, Q_mat);
@@ -63,13 +81,13 @@ Rcpp::List SimulateC(const int& nsim,
   for (int i = 0; i < N; i++) {
 
     // Get system matrices of current timepoint
-    if (Z_tv && i > 0) {
+    if (Z_tv && !eta_only && i > 0) {
       Z_mat = Z.slice(i);
     }
-    if (T_tv && i > 0) {
+    if (T_tv && !eta_only && i > 0) {
       T_mat = T.slice(i);
     }
-    if (R_tv && i > 0) {
+    if (R_tv && !eta_only && i > 0) {
       R_mat = R.slice(i);
     }
     if (Q_tv && i > 0) {
@@ -79,7 +97,9 @@ Rcpp::List SimulateC(const int& nsim,
     }
 
     // Simulated y component
-    y.row(i) = Z_mat * a_temp;
+    if (!eta_only) {
+      y.row(i) = Z_mat * a_temp;
+    }
 
     // Draw random numbers following the standard normal distribution
     draw = Rcpp::rnorm(total);
@@ -94,7 +114,7 @@ Rcpp::List SimulateC(const int& nsim,
     eta.row(i) = eta_temp;
 
     // Simulated state
-    if (i < N_min1) {
+    if (!eta_only && i < N_min1) {
       a_temp = T_mat * a_temp + R_mat * eta_temp;
       a_cube.row(i + 1) = a_temp;
     }